Revert "[pybind] Moves a historically flaky RenderEngine test into its own test (#14724)"
This reverts commit 5a7203e357756229e5fc3508c5d68684750088e7.
